refactor: extract SynthDriver into adapters/nvda/, decouple backend bootstrap
Move the driver classes (SynthDriver, SpeechTask, BreakTask,
IndexReachedTask, DoneSpeakingTask, process_speech) out of the
top-level package __init__.py into adapters/nvda/synth_driver.py.

Bootstrap (starting the gRPC backend and loading voices) is now a
module-level _bootstrap_backend() function that tests can replace
wholesale via monkeypatch, since NVDA constructs SynthDriver() with
zero arguments and the backend can't be a constructor parameter.
__init__ narrows its typed catch to BackendUnavailableError, keeping
one final broad except-Exception as a last-resort safety net so an
unanticipated bootstrap bug can't take down NVDA's whole speech
subsystem.

The top-level __init__.py becomes a thin re-export shim; NVDA's
synthDriverHandler discovers the driver by importing this package and
reading SynthDriver at the top level, so the re-export is
load-bearing.

tests/test_synth_driver.py now loads the driver from its new location
and injects FakeTTSBackend by monkeypatching _bootstrap_backend
instead of mocking grpc_client.
