+++ addon/synthDrivers/dengjen_neural_voices/adapters/nvda/__init__.py
@@ -1,0 +1,1 @@
+# coding: utf-8
